+++ extensions/littlehorse-quarkus/runtime/src/main/java/io/littlehorse/quarkus/adapter/LHTypeAdapter.java
@@ -1,0 +1,17 @@
+package io.littlehorse.quarkus.adapter;
+
+import io.littlehorse.sdk.common.proto.VariableType;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LHTypeAdapter {
+
+    VariableType variableType();
+
+    Class<?> adaptedType();
+}
